refactor: break up voting component
