+++ app/components/_shared/vote-loading-state/vote-loading-state.tsx
@@ -1,0 +1,17 @@
+import { Card, Loader } from "..";
+import VotingCardTitle from "../vote-voting-card-title/voting-card-title";
+
+const LoadingState = () => {
+  return (
+    <Card>
+      <VotingCardTitle />
+      <div className="flex flex-col gap-x3 mt-x2 text-center">
+        <br />
+        <Loader isCenter />
+        <br />
+      </div>
+    </Card>
+  );
+};
+
+export default LoadingState;
