+++ app/components/_shared/vote-voting-card-title/voting-card-title.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Card } from "@/app/components/_shared";
+const VotingCardTitle = ({
+  children = "Cast Votes",
+}: React.PropsWithChildren) => {
+  return (
+    <Card.Header className="text-center">
+      <h2 className="text-[2rem] leading-[2rem] font-fg font-medium">
+        {children}
+      </h2>
+    </Card.Header>
+  );
+};
+
+export default VotingCardTitle;
