+++ app/components/vote-cast-vote/cast-vote.tsx
@@ -1,0 +1,167 @@
+import { useAccount } from "wagmi";
+
+import { UserRejectedRequestError } from "viem";
+
+import { Button, Card } from "@/app/components/_shared";
+import { ChevronIcon } from "@/app/components/_icons";
+import BlockExplorerLink from "@/app/components/_shared/block-explorer-link/block-explorer-link.component";
+
+import { Proposal } from "@/app/graphql";
+import { useChainState } from "@/app/providers/chainState.provider";
+import useVoteReceipt from "@/app/hooks/useVoteReceipt";
+import useCastVote from "@/app/hooks/useCastVote";
+
+import ErrorHelper from "@/app/helpers/error.helper";
+import { SuccessIcon } from "@/app/components/_icons/success-icon";
+import VotingCardTitle from "../_shared/vote-voting-card-title/voting-card-title";
+import DisconnectedState from "./disconnected-state";
+import LoadingState from "../_shared/vote-loading-state/vote-loading-state";
+import VotingButtons from "./voting-buttons";
+import LockedBalance from "./locked-balance";
+
+import HasVoted from "./has-voted";
+import VoteConfirmation from "./vote-confirmation";
+
+export const VOTE_TYPES = {
+  Against: 0,
+  For: 1,
+  Abstain: 2,
+} as { [key: string]: number };
+
+export const REVERSE_VOTE_TYPE_MAP = {
+  [VOTE_TYPES.For]: "For",
+  [VOTE_TYPES.Against]: "Against",
+  [VOTE_TYPES.Abstain]: "Abstain",
+} as const;
+
+const CastVote = ({ proposalId }: { proposalId: Proposal["proposalId"] }) => {
+  const { address, isConnected, isConnecting, isDisconnected } = useAccount();
+  const tokens = useChainState((s) => s.tokens);
+  const chainStateReady = useChainState((s) => s.ready);
+  const { data: voteReceipt, isLoading: isHasVotedStatusLoading } =
+    useVoteReceipt({
+      proposalId,
+      address,
+    });
+
+  const vote = useCastVote();
+
+  const hasEnoughLockedMentoToVote = tokens.veMento.balance > 0;
+  const isInitializing =
+    isConnecting ||
+    (isConnected && !chainStateReady) ||
+    isHasVotedStatusLoading;
+
+  const handleCastVote = (voteType: number) => {
+    try {
+      vote.castVote(proposalId, voteType);
+    } catch (error) {
+      // TODO: sentrify me
+      console.log(error);
+    }
+  };
+
+  if (isInitializing) {
+    return <LoadingState />;
+  }
+
+  if (isDisconnected) {
+    return <DisconnectedState />;
+  }
+
+  if (voteReceipt && voteReceipt.hasVoted) {
+    return <HasVoted voteType={voteReceipt.support} />;
+  }
+
+  if (!hasEnoughLockedMentoToVote) {
+    return <DirectToLockMento />;
+  }
+
+  if (vote.isAwaitingUserSignature) {
+    return (
+      <VoteConfirmation
+        voteType={vote?.variables?.args?.[1] as number}
+        proposalId={proposalId}
+      />
+    );
+  }
+
+  if (vote.isConfirming) {
+    return (
+      <Card>
+        <VotingCardTitle />
+        <div className="flex flex-col gap-x3 mt-x2 text-center">
+          <span className="text-md">Vote submitted</span>
+          <SuccessIcon className="w-20 h-20 mx-auto" />
+          <span className="text-[#A8A8A8] dark:text-[#AAB3B6] text-sm">
+            Your vote is being processed
+          </span>
+          {vote.hash && (
+            <BlockExplorerLink type="tx" item={vote.hash}>
+              View on explorer
+            </BlockExplorerLink>
+          )}
+        </div>
+      </Card>
+    );
+  }
+
+  if (vote.isConfirmed) {
+    return (
+      <Card>
+        <VotingCardTitle />
+        <div className="flex flex-col gap-x3 mt-x2 text-center">
+          <span className="text-md">Vote success</span>
+          <SuccessIcon className="w-20 h-20 mx-auto" />
+          {vote.hash && (
+            <BlockExplorerLink type="tx" item={vote.hash}>
+              View on explorer
+            </BlockExplorerLink>
+          )}
+        </div>
+      </Card>
+    );
+  }
+
+  return (
+    <Card>
+      <VotingCardTitle />
+      <div className="flex flex-col gap-x5 mt-x3">
+        <LockedBalance />
+        <div className="flex flex-col gap-2">
+          <VotingButtons onSubmit={handleCastVote} />
+          {vote.error && <VotingError error={vote.error} />}
+        </div>
+      </div>
+    </Card>
+  );
+};
+
+const VotingError = ({ error }: { error: Error }) => {
+  if (error instanceof UserRejectedRequestError) {
+    return null;
+  }
+  return (
+    <div className="text-light-red flex flex-col gap-1 items-center justify-center text-sm w-full">
+      <span>Error casting vote</span>
+      <span>{ErrorHelper.processWagmiErrorMessage(error)}</span>
+    </div>
+  );
+};
+
+const DirectToLockMento = () => {
+  return (
+    <Card>
+      <VotingCardTitle />
+      <div className="flex flex-col gap-x5">
+        <LockedBalance />
+        <span>You need to lock your MENTO to vote</span>
+        <Button href="/my-voting-power" theme="primary">
+          Lock Mento <ChevronIcon direction="right" />
+        </Button>
+      </div>
+    </Card>
+  );
+};
+
+export default CastVote;
