+++ app/components/vote-inactive-proposal-status-message/inactive-proposal-status-message.tsx
